+++ changelog.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keep-a-changelog],
 and this project adheres to [Semantic Versioning][semver].
 
-<!-- ## [X.Y.Z] - YYYY-MM-DD
+<!-- ## [X.Y.Z]
+_YYYY-MM-DD_
 
 ### Added
 
@@ -33,8 +34,31 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
--   _TBD_
+- _TBD_
 
-[unreleased]: https://github.com/bradgarropy/remix-starter/compare/v3.0.0...HEAD
+## [4.0.0][4.0.0]
+
+_2024-12-10_
+
+- Upgrade to [React 19][react-19]
+- Upgrade to [Vite 6][vite-6]
+- Adopt Remix future flags
+    - [`v3_fetcherPersist`][v3-fetcherPersist]
+    - [`v3_relativeSplatPath`][v3-relativeSplatPath]
+    - [`v3_throwAbortReason`][v3-throwAbortReason]
+    - [`v3_lazyRouteDiscovery`][v3-lazyRouteDiscovery]
+    - [`v3_singleFetch`][v3-singleFetch]
+- Remove [`@remix-run/eslint-config`][remix-run-eslint-config]
+
+[unreleased]: https://github.com/bradgarropy/remix-starter/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/bradgarropy/remix-starter/releases/tag/v4.0.0
 [keep-a-changelog]: https://keepachangelog.com
 [semver]: https://semver.org
+[react-19]: https://react.dev/blog/2024/12/05/react-19
+[vite-6]: https://vite.dev/blog/announcing-vite6
+[v3-fetcherPersist]: https://remix.run/docs/en/main/start/future-flags#v3_fetcherpersist
+[v3-relativeSplatPath]: https://remix.run/docs/en/main/start/future-flags#v3_relativesplatpath
+[v3-throwAbortReason]: https://remix.run/docs/en/main/start/future-flags#v3_throwabortreason
+[v3-lazyRouteDiscovery]: https://remix.run/docs/en/main/start/future-flags#v3_lazyroutediscovery
+[v3-singleFetch]: https://remix.run/docs/en/main/start/future-flags#v3_singlefetch
+[remix-run-eslint-config]: https://remix.run/docs/en/main/start/future-flags#remix-runeslint-config